requirements.txt
# Core application dependencies
streamlit>=1.28.0
duckdb==0.10.0
confluent-kafka==2.10.1
pyspark==3.3.2
pandas>=2.0.0
pydeck>=0.8.0
streamlit-autorefresh>=1.0.0
delta-spark==2.3.0
python-dotenv>=1.0.0

# Testing dependencies
pytest>=8.0.0
black>=24.0.0
flake8>=7.0.0
plotly>=6.0.0
pytest-mock>=3.12.0
pytest-cov>=4.1.0

# Airflow 3.0.0 and compatible dependencies
apache-airflow==3.0.0
apache-airflow-providers-docker>=3.7.0

# System dependencies
psutil>=6.1.0
